fix selector bug and compose baseSelector
